wpt_babel/content-security-policy/inheritance/iframe-all-local-schemes.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js">
</script>
<script src="/resources/testharnessreport.js">
</script>
<meta content="img-src 'none'" http-equiv="Content-Security-Policy"/>
<body>
 <script>
  "use strict";

function wait_for_error_from_frame(frame, test) {
  window.addEventListener('message', test.step_func(function (e) {
    if (e.source != frame.contentWindow) return;
    assert_equals(e.data, "error");
    frame.remove();
    test.done();
  }));
}
function wait_for_error_from_window(opened_window, test) {
  window.addEventListener('message', test.step_func(function (e) {
    if (e.source != opened_window) return;
    assert_equals(e.data, "error");
    opened_window.close();
    test.done();
  }));
}
async_test(function (t) {
  var i = document.createElement('iframe');
  document.body.appendChild(i);
  var img = document.createElement('img');
  img.onerror = t.step_func_done(function (_) {
    return i.remove();
  });
  img.onload = t.unreached_func();
  i.contentDocument.body.appendChild(img);
  img.src = "{{location[server]}}/images/red-16x16.png";
}, "<iframe>'s about:blank inherits policy.");
async_test(function (t) {
  var w = window.open("about:blank");
  var then = t.step_func(function () {
    then = function then() {};
    var img = w.document.createElement('img');
    img.onerror = t.step_func_done(function (_) {
      return w.close();
    });
    img.onload = t.unreached_func();
    w.document.body.appendChild(img);
    img.src = "{{location[server]}}/images/red-16x16.png";
  });

  // There are now interoperable way to wait for the initial about:blank
  // document to load. Chrome loads it synchronously, hence we can't wait for
  // w.onload. On the other side Firefox loads the initial empty document
  // later and we can wait for the onload event.
  w.onload = then;
  setTimeout(then, 200);

  // Navigations to about:blank happens synchronously. There is no need to
  // wait for the document to load.
}, "window about:blank inherits policy.");
async_test(function (t) {
  var i = document.createElement('iframe');
  i.srcdoc = "\n      <img src='{{location[server]}}/images/red-16x16.png'\n        onload='window.top.postMessage(\"load\", \"*\");'\n        onerror='window.top.postMessage(\"error\", \"*\");'\n      >\n    ";
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe srcdoc>'s inherits policy.");
async_test(function (t) {
  var i = document.createElement('iframe');
  var b = new Blob(["\n        <img src='{{location[server]}}/images/red-16x16.png'\n          onload='window.top.postMessage(\"load\", \"*\");'\n          onerror='window.top.postMessage(\"error\", \"*\");'\n        >\n      "], {
    type: "text/html"
  });
  i.src = URL.createObjectURL(b);
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe src='blob:...'>'s inherits policy.");
async_test(function (t) {
  var b = new Blob(["\n        <img src='{{location[server]}}/images/red-16x16.png'\n          onload='window.opener.postMessage(\"load\", \"*\");'\n          onerror='window.opener.postMessage(\"error\", \"*\");'\n        >\n      "], {
    type: "text/html"
  });
  var url = URL.createObjectURL(b);
  var w = window.open(url);
  wait_for_error_from_window(w, t);
}, "window url='blob:...' inherits policy.");
async_test(function (t) {
  var i = document.createElement('iframe');
  i.src = "data:text/html,<img src='{{location[server]}}/images/red-16x16.png'\n      onload='window.top.postMessage(\"load\", \"*\");'\n      onerror='window.top.postMessage(\"error\", \"*\");'\n    >";
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe src='data:...'>'s inherits policy.");

// Opening a window toward a data-url isn't allowed anymore. Hence, it can't
// be tested.

async_test(function (t) {
  var i = document.createElement('iframe');
  i.src = "javascript:\"<img src='{{location[server]}}/images/red-16x16.png'\n      onload='window.top.postMessage(\\\"load\\\", \\\"*\\\");'\n      onerror='window.top.postMessage(\\\"error\\\", \\\"*\\\");'\n    >\"";
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe src='javascript:...'>'s inherits policy (static <img> is blocked)");
async_test(function (t) {
  var url = "javascript:\"<img src='{{location[server]}}/images/red-16x16.png'\n      onload='window.opener.postMessage(\\\"load\\\", \\\"*\\\");'\n      onerror='window.opener.postMessage(\\\"error\\\", \\\"*\\\");'\n    >\"";
  var w = window.open(url);
  wait_for_error_from_window(w, t);
}, "window url='javascript:...'>'s inherits policy (static <img> is blocked)");

// Same as the previous javascript-URL test, but instead of loading the <img>
// from the new document, this one is created from the initial empty document,
// while evaluating the javascript-url.
// See https://crbug.com/1064676
async_test(function (t) {
  var url = "javascript:\n      let img = document.createElement('img');\n      img.onload = () => window.top.postMessage('load', '*');\n      img.onerror = () => window.top.postMessage('error', '*');\n      img.src = '{{location[server]}}/images/red-16x16.png';\n      document.body.appendChild(img);\n    ";
  var i = document.createElement('iframe');
  i.src = encodeURI(url.replace(/\n/g, ""));
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe src='javascript:...'>'s inherits policy (dynamically inserted <img> is blocked)");
async_test(function (t) {
  var i = document.createElement('iframe');
  var b = new Blob(["\n        <img src='{{location[server]}}/images/red-16x16.png'\n          onload='window.top.postMessage(\"load\", \"*\");'\n          onerror='window.top.postMessage(\"error\", \"*\");'\n        >\n      "], {
    type: "text/html"
  });
  i.src = URL.createObjectURL(b);
  i.sandbox = 'allow-scripts';
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe sandbox src='blob:...'>'s inherits policy. (opaque origin sandbox)");
 </script>
</body>
